Saved a copy of all the classes with the comments to notepad++ for reference and amended the cclasses in the project to just have my code in it so it is easier to read and for my head to follow the flow Went back to pet,cat and dog adding getters , setter, equals and toString. Started to see a couple of errors coming in at this point wasn't verifying it was working as I was going so need to stop and review at this point. equals I just added as spec said I don't fully understand this part(Going back to review some lecture notes on this)
Will look to go back over from the start as I went a long and validate as I go that it is correct before moving on any further.

As I am going back over the Pet,cat and dog classes I am reviewing what I added. I was unsure originally when working through the constructore and setters how to impliment the utils into them.
I have reviewed the lectures ands was able to impliment correctly and looks good before moving on.

Amended public to private from initial boilerplate fields in pet , dog , cat as I want to have proper encapsulation
made changes to getters and setters setters should be void as performing an update doesn't need to return anything
Also updated next Id and Id. Addeed nextId as a static int and it is added to the constructor when it is called but not added as aparameter as it will be added automatically.

Added the overide calculateWeeklyFee with a basic return of 0 in cat and dog for now so implementation there so java wont add it

Added the methods for checkin,checkout and NumofdayinKennel.
made use of the utilities validrange utility and added the .length -1 check as array starts at 0 to get correct range and insure no arraayoutofbound errors.
Added a counter for the numdaysinkennel with an enhanced forloop to run through it
there is also a parameter in the constructor for Dog int age which i think is a typo as it is not in the UML.
maybe it is there as one we can add it but left it out for now

 Looking at adding the details for the calculateweekly rate for the dog class complier advised DANGEROUS_DAILY_RATE and NONDANGEROUS_DAILY_RATE
 might be final. changed them to final as they would not be changed elsewhere an if rates went up they would be changed her an kept as final.
 As no other part of the application decided if therer is other charges to be added here
 removed the setters for the dangerous breeds as no need to set them as they are final an won't be changed I think that would be best practice to do.

 Added an If statement to get the daily rate for the dangerous breed or not and multiplied it against the days in kennel and returned as a double

 Used a simialr approach to the cat class for the weekly charge an If/else statement based on if it is indoor cat or not an then multiplied by numofdays